Make colours independant of the Gtk theme
In some themes, the colour obtained from gtk_style_context_get_color() and used for the illuminated area is darker than the background.

This is counter intuitive.

Since we're working with a hardcoded backgroound image, I think we can hardcode the colour for the illuminated area. These values were obtained from the Adwaita-dark theme and seem to give a reasonable result - but anything lighter than the background would do.